Simplify injection regex and attempt to match indentation
Instead of matching then entire <head>...</head> block, match the
whitespace before </head> and then use that to indent our injected
script correctly.

This has the advantage that we don't need to capture the entire content
of <head>...</head> and also won't break if <head> has attributes or
other abnormalities.

It will break, however, if the file contains the string "</head>"
anywhere before the actual </head> closing tag, however IMO this is far
less likely than the <head> tag having attributes.

Also use %r{} instead of // for regex notation so that we don't have to
escape the / in </head>.